requirements.txt
# Runtime dependencies for BNB Profit Harvester Bot
python-binance==1.0.32
python-dotenv==1.0.0
requests==2.32.5
aiohttp==3.13.1
websockets==15.0.1
pycryptodome==3.23.0
pytz==2025.2
regex==2025.10.23
# Optional but present in venv; harmless if unused
numpy==2.3.4
pandas==2.3.3

